Split Name::Validation tests out of name_test.rb (#4708)
Moves the 8 tests exercising Name::Validation (user_presence,
text_name/author length limits, author character/ending rules,
search_name_indistinct, search_name presence) into
test/models/name/validation_test.rb, matching
app/models/name/validation.rb.

test_version_records_editor_not_creator stays behind - it exercises
acts_as_versioned/VersionedByCurrentUser on the base Name model, not
a splittable concern module. test_show_includes_omits_namings_and_observations
and test_merge_includes_preloads_namings_and_observations also stay -
they exercise Name::Scopes and will move with the future Scopes split.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
